+++ zcash_client_sqlite/src/wallet/transparent.rs
@@ -131,7 +131,8 @@ pub(crate) fn get_transparent_receivers<P: consensus::Parameters>(
             key_scope,
             transparent_child_index,
             imported_transparent_receiver_pubkey,
-            exposed_at_height
+            exposed_at_height,
+            transparent_receiver_next_check_time
          FROM addresses
          JOIN accounts ON accounts.id = addresses.account_id
          WHERE accounts.uuid = :account_uuid
@@ -167,6 +168,11 @@ pub(crate) fn get_transparent_receivers<P: consensus::Parameters>(
                 ExposedAt::Height(BlockHeight::from(h))
             });
 
+        let next_check_time = row
+            .get::<_, Option<i64>>("transparent_receiver_next_check_time")?
+            .map(decode_epoch_seconds)
+            .transpose()?;
+
         if let Some(taddr) = taddr {
             let metadata = match key_scope {
                 #[cfg(feature = "transparent-key-import")]
@@ -190,7 +196,7 @@ pub(crate) fn get_transparent_receivers<P: consensus::Parameters>(
                     let pubkey = PublicKey::from_bytes(pubkey_bytes).map_err(|e| {
                         SqliteClientError::CorruptedData(format!("Invalid public key: {}", e))
                     })?;
-                    TransparentAddressMetadata::standalone(pubkey, exposed_at)
+                    TransparentAddressMetadata::standalone(pubkey, exposed_at, next_check_time)
                 }
                 derived => {
                     let scope_opt = <Option<TransparentKeyScope>>::from(derived);
@@ -212,6 +218,7 @@ pub(crate) fn get_transparent_receivers<P: consensus::Parameters>(
                             )),
                         )?,
                         exposed_at,
+                        next_check_time,
                     )
                 }
             };
@@ -421,6 +428,7 @@ pub(crate) fn select_addrs_to_reserve<P: consensus::Parameters>(
                                         scope,
                                         i,
                                         ExposedAt::Unexposed,
+                                        None,
                                     ))
                                 },
                             )?,
@@ -1505,7 +1513,12 @@ pub(crate) fn get_transparent_address_metadata<P: consensus::Parameters>(
     let address_str = address.encode(params);
     let addr_meta = conn
         .query_row(
-            "SELECT diversifier_index_be, key_scope, imported_transparent_receiver_pubkey, exposed_at_height
+            "SELECT
+                diversifier_index_be,
+                key_scope,
+                imported_transparent_receiver_pubkey,
+                exposed_at_height,
+                transparent_receiver_next_check_time
              FROM addresses
              JOIN accounts ON addresses.account_id = accounts.id
              WHERE accounts.uuid = :account_uuid
@@ -1518,6 +1531,12 @@ pub(crate) fn get_transparent_address_metadata<P: consensus::Parameters>(
                     |h| ExposedAt::Height(BlockHeight::from(h))
                 );
 
+                let next_check_time = row
+                    .get::<_, Option<i64>>("transparent_receiver_next_check_time")?
+                    .map(decode_epoch_seconds)
+                    .transpose()
+                    .map_err(|e| rusqlite::Error::ToSqlConversionFailure(Box::new(e)))?;
+
                 Ok(KeyScope::decode(scope_code).and_then(|key_scope| {
                     let address_index = address_index_from_diversifier_index_be(row.get("diversifier_index_be")?)?;
                     match <Option<TransparentKeyScope>>::from(key_scope).zip(address_index) {
@@ -1525,7 +1544,8 @@ pub(crate) fn get_transparent_address_metadata<P: consensus::Parameters>(
                             Ok(TransparentAddressMetadata::derived(
                                 scope,
                                 address_index,
-                                exposed_at
+                                exposed_at,
+                                next_check_time
                             ))
                         }
                         None => {
@@ -1543,7 +1563,7 @@ pub(crate) fn get_transparent_address_metadata<P: consensus::Parameters>(
                                         )
                                     })?;
                                     let pubkey = secp256k1::PublicKey::from_bytes(pubkey_bytes)?;
-                                    Ok(TransparentAddressMetadata::standalone(pubkey, exposed_at))
+                                    Ok(TransparentAddressMetadata::standalone(pubkey, exposed_at, next_check_time))
                                 };
 
                                 addr_meta
@@ -1570,6 +1590,7 @@ pub(crate) fn get_transparent_address_metadata<P: consensus::Parameters>(
                 Scope::External.into(),
                 address_index,
                 ExposedAt::Unknown,
+                None,
             );
             return Ok(Some(metadata));
         }
@@ -1916,6 +1937,7 @@ mod tests {
                     ephemeral::metadata(
                         NonHardenedChildIndex::from_index(i).unwrap(),
                         ExposedAt::Unexposed,
+                        None,
                     )
                 })
                 .collect();